+++ examples/src/exn.ml
@@ -1,0 +1,2 @@
+type bar = int
+exception Foo of bar