+++ examples/src/exn.mli
@@ -1,0 +1,2 @@
+type bar
+exception Foo of bar