+++ theories/Exn.v
@@ -1,0 +1,11 @@
+From Coq Require Extraction.
+
+Axiom exn : Type.
+
+Class Exn (e : Type) := { to_exn : e -> exn
+                        ; of_exn : exn -> option e
+                        }.
+
+Module ExnExtraction.
+  Extract Constant exn => "exn".
+End ExnExtraction.